Abundance measured in number of stems, basal area or above ground biomass.

Abundance measured in number of stems, basal area or above ground biomass.

abundance(censdata, type = "abund", alivecode = c("A"), mindbh = NULL,
  dbhunit = "mm", split1 = NULL, split2 = NULL)

abundanceperquad(censdata, mindbh = 10, plotdim = c(1000, 500),
  gridsize = 100, type = "abund", dbhunit = "mm")

abundance_match_census_habitat(censdata, habitats)

Arguments

censdata
A table of plot census data.

type
Either 'abund' (default) for abundance, 'ba' for basal area, or 'agb'.

alivecode
character, codes of the variable `status` that indicate the tree is alive. The default 'A' is the standard CTFS designation for living trees or stems

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

dbhunit
'cm' or 'mm', only used for basal area.

split1, split2
A vector of categories, one per individual. See details.

plotdim
The x and y dimensions of the plot.

gridsize
Side of the square quadrate.

habitats
Data frame giving the habitat designation for each 20x20 quadrat.

Value

abundance() and abundanceperquad() return a named list of two elements:

either abund or ba

meandate (mean measurement date for all individuals in each category).

abundance_match_census_habitat(): A dataframe giving the abundance per quadrat.

Details

abundance() calculates total abundance or basal area, dividing data with 1 or 2 categorical variables passed to split1 and split2. The length of split1 and split2 must match exactly the number of rows in censdata (so one per tree or stem). split1 should be the one with the most categories (for instances, split1 = species, split2 = dbhcategory).

Each element of the returned list is an array of one or two dimensions, depending on how many split variables were submitted: each of the dimensions of the array handles one of the categorical variables.

For basal area, pass a stem table to censdata; For abundance, use either the stem or full table to count stems or trees, respectively.

abundanceperquadrat() wraps abundance() to find abundance, basal area, or agb of every species per square quadrat of any size.

If the plot dimensions is not an exact multiple of the quadrat size, then a strip at the upper edge of the plot (north and east if plot is on cardinal directions) is omitted. For example, if gridsize = 40 and plotdim = 500, then there are an extra 20 meters at the upper boundary omitted from the calculations.

The array of abundances per quadrat is useful for similarity, counting species and stems per quadrat, etc.

abundance_match_census_habitat() wraps abundanceperquad() ensuring consistent gridsize and plotdim via extract_gridsize() and extract_plotdim().

Acknowledgement

Thanks to Gabriel Arellano for suggesting that the match plot dimensions between habitat data and census data should be enforced programatically.

See also

ba, ba_sum

Examples

# Small subsets of data for examples
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

tree_mini <- dplyr::sample_n(bci::bci12full5, 1000)
stem_mini <- bci::bci12stem5[1:1000, ]

(tree_abundance <- abundance(tree_mini, mindbh = 10))

#> $abund
#>     all
#> all 511
#> 
#> $meandate
#>          all
#> all 14746.64
#> 

(stem_abundance <- abundance(stem_mini, mindbh = 10))

#> $abund
#>     all
#> all 423
#> 
#> $meandate
#>          all
#> all 14867.36
#> 

stem_basal_area_by_sp <- abundance(
  stem_mini,
  type = "ba",
  mindbh = 10,
  split1 = stem_mini$sp
)
lapply(stem_basal_area_by_sp, head)

#> $ba
#>               all
#> aegipa 0.00000000
#> alchco 0.17657636
#> allops 0.06326225
#> alsebl 3.49413040
#> amaico 0.00000000
#> anacex 8.82765073
#> 
#> $meandate
#>             all
#> aegipa       NA
#> alchco 14872.67
#> allops 14871.00
#> alsebl 14869.03
#> amaico       NA
#> anacex 14872.60
#> 

tree_n <- abundanceperquad(
  tree_mini,
  plotdim = c(1000, 500),
  gridsize = 100,
  type = "abund"
)
str(tree_n$abund, list.len = 10)

#> 'data.frame':	143 obs. of  50 variables:
#>  $ 1 : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 2 : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 3 : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 4 : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ 5 : num  0 0 0 0 0 1 0 1 0 0 ...
#>  $ 6 : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 7 : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 8 : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ 9 : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 10: num  0 0 0 0 1 0 0 0 0 0 ...
#>   [list output truncated]

# Matching `gridsize` and `plotdim` between census data and habitat data
abund <- abundance_match_census_habitat(tree_mini, bci::bci_habitat)
str(abund, list.len = 6)

#> 'data.frame':	143 obs. of  1250 variables:
#>  $ 1   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 2   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 3   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 4   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 5   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ 6   : num  0 0 0 0 0 0 0 0 0 0 ...
#>   [list output truncated]