Create tables of abundance and basal area by year.

These functions first automatically pick the main stem (see ?fgeo.tool::pick_main_stem()). Then, f or each species and each (round mean) year of measurement, abundance_byyr() counts the number of trees, and basal_area_byyr() calculates the total basal area.

abundance_byyr(vft, ...)

basal_area_byyr(vft, ...)

Arguments

vft
A ForestGEO-like dataframe; particularly a ViewFullTable. As such, it should contain columns PlotName, CensusID, TreeID, StemID, Status, DBH, Genus, SpeciesName, ExactDate, PlotCensusNumber, Family, Tag, and HOM. ExactDate should contain dates from 1980-01-01 to the present day in the format yyyy-mm-dd.

...
Expressions to pick main stems of a specific dbh range.

Value

A dataframe.

Details

You don't need to pick stems by status before feeding data to these functions. Doing so may make your code more readable but it should not affect the result. This is because the expressions passed to ... pick data by dbh and exclude missing the dbh values associated to non-alive stems, including dead, missing, and gone.

See also

Examples

library(fgeo.tool)

#> 
#> Attaching package: 'fgeo.tool'

#> The following object is masked from 'package:stats':
#> 
#>     filter

vft <- example_byyr
vft

#> # A tibble: 8 x 13
#>   PlotName CensusID TreeID StemID Status   DBH Genus SpeciesName ExactDate 
#>   <chr>       <int>  <int>  <dbl> <chr>  <int> <chr> <chr>       <date>    
#> 1 luq             1      1    1.1 alive     10 Gn    spp         2001-01-01
#> 2 luq             1      1    1.2 dead      NA Gn    spp         2001-01-01
#> 3 luq             1      2    2.1 alive     20 Gn    spp         2001-01-01
#> 4 luq             1      2    2.2 alive     30 Gn    spp         2001-01-01
#> 5 luq             2      1    1.1 alive     20 Gn    spp         2002-01-01
#> 6 luq             2      1    1.2 gone      NA Gn    spp         2002-01-01
#> 7 luq             2      2    2.1 dead      NA Gn    spp         2002-01-01
#> 8 luq             2      2    2.2 dead      NA Gn    spp         2002-01-01
#> # ... with 4 more variables: PlotCensusNumber <int>, Family <chr>, Tag <int>,
#> #   HOM <int>

abundance_byyr(vft, DBH >= 10, DBH < 20)

#> # A tibble: 1 x 3
#>   species Family `2001`
#>   <chr>   <chr>   <dbl>
#> 1 Gn spp  f           1

abundance_byyr(vft, DBH >= 10)

#> # A tibble: 1 x 4
#>   species Family `2001` `2002`
#>   <chr>   <chr>   <dbl>  <dbl>
#> 1 Gn spp  f           2      1

basal <- basal_area_byyr(vft, DBH >= 10)
# Convert units and standardize by plot size in hectares
years <- c("2001", "2002")
in_he <- convert_unit_at(basal, .at = years, from = "mm2", to = "hectare")
standardize_at(in_he, .at = years, denominator = 50)

#> # A tibble: 1 x 4
#>   species Family        `2001`   `2002`
#>   <chr>   <chr>          <dbl>    <dbl>
#> 1 Gn spp  f      0.00000000157 6.28e-10