Count individuals.

Count individuals.

abundance_n(x, group_by = c("quadrat", "sp"), only_alive = TRUE)

Arguments

x
Census.

group_by
Sting giving the variables to group by.

only_alive
If TRUE, filter only alive individuals.

Value

As per base::table().

See also

base::table().

Examples

stem <- forestr::bci12t7mini
(n <- abundance_n(stem))

#>        sp
#> quadrat acaldi alibed alsebl anaxpa andiin annoac apeime aspicr beilpe brosal
#>    0000      0      0      0      0      0      0      0      0      0      0
#>    0002      0      0      0      0      0      0      0      0      0      0
#>    0006      0      0      0      0      0      0      0      0      0      0
#>    0009      0      0      0      0      0      0      0      0      0      0
#>    0020      0      0      0      0      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      0
#>    0100      0      0      0      0      0      0      0      0      0      0
#>    0101      0      0      0      0      0      0      0      0      0      0
#>    0102      0      0      0      0      0      0      0      0      0      0
#>        sp
#> quadrat calolo cappfr caseac casesy cassel cecrin cha2sc chr1ec chr2ar coccma
#>    0000      0      0      0      0      0      0      0      0      0      0
#>    0002      0      0      0      0      0      0      0      0      0      0
#>    0006      0      0      0      0      0      0      0      0      0      0
#>    0009      0      0      0      0      0      0      0      0      0      0
#>    0020      0      0      0      0      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      0
#>    0100      0      0      0      0      0      0      0      0      0      0
#>    0101      0      0      0      0      0      0      0      0      0      0
#>    0102      0      0      0      0      0      0      0      0      0      0
#>        sp
#> quadrat cordal cordbi cordla cou2cu cupasy des2pa dio2ar drypst ery2ma eugeco
#>    0000      0      0      0      0      0      0      0      0      0      0
#>    0002      0      0      0      0      0      0      0      0      0      0
#>    0006      0      0      0      0      0      0      0      0      0      0
#>    0009      0      0      0      0      0      0      0      0      0      0
#>    0020      0      0      0      0      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      0
#>    0100      0      0      0      0      0      1      0      0      0      0
#>    0101      0      0      0      0      0      0      0      0      0      0
#>    0102      0      0      0      0      0      0      0      0      0      0
#>        sp
#> quadrat eugega eugene eugeoe faraoc gar2in guapst guargu guarsp guatdu gustsu
#>    0000      0      0      0      0      0      0      0      0      0      0
#>    0002      0      0      0      0      0      0      0      0      0      0
#>    0006      0      0      0      0      0      0      0      0      0      0
#>    0009      0      0      0      1      0      0      0      0      0      0
#>    0020      0      0      0      0      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      0
#>    0100      0      0      0      0      0      0      0      0      0      0
#>    0101      0      0      0      0      0      0      0      0      0      0
#>    0102      0      0      0      0      0      0      0      0      0      0
#>        sp
#> quadrat hampap heisco herrpu hirttr hybapr ingafa ingago ingama ingami ingapu
#>    0000      0      0      0      0      0      0      0      0      0      0
#>    0002      0      0      0      0      1      0      0      0      0      0
#>    0006      0      0      0      0      1      0      0      0      0      0
#>    0009      0      0      0      0      1      0      0      0      0      0
#>    0020      0      0      0      0      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      0
#>    0100      0      0      0      0      0      0      0      0      0      0
#>    0101      0      0      0      0      0      1      0      0      0      0
#>    0102      0      0      0      0      1      0      0      0      0      1
#>        sp
#> quadrat ingaqu ingas1 ingaum laciag laetth licapl luehse malmsp maquco micoar
#>    0000      0      0      0      0      0      0      0      0      0      0
#>    0002      0      0      0      0      0      0      0      0      0      0
#>    0006      0      0      0      0      0      0      0      0      0      0
#>    0009      0      0      0      0      0      0      0      0      0      0
#>    0020      0      0      0      0      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      0
#>    0100      0      0      0      0      0      0      0      0      0      0
#>    0101      0      0      0      0      0      0      0      0      0      0
#>    0102      0      0      0      0      0      0      0      0      0      0
#>        sp
#> quadrat micone mourmy nectgl ocotob ocotpu ocotwh oenoma ormocr ouralu paligu
#>    0000      0      0      0      0      0      0      0      0      0      0
#>    0002      0      0      0      0      0      0      0      0      0      0
#>    0006      0      0      0      0      0      0      0      0      0      0
#>    0009      0      0      0      0      0      0      0      0      0      0
#>    0020      0      0      0      0      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      0
#>    0100      0      0      0      0      0      0      0      0      0      0
#>    0101      0      1      0      0      0      0      0      0      0      0
#>    0102      0      0      0      0      0      0      0      0      0      0
#>        sp
#> quadrat pentma picrla pla1pi poular poutre pri2co protco protpa protte pse1se
#>    0000      0      0      0      0      0      0      0      0      0      0
#>    0002      0      0      0      0      0      0      0      0      0      0
#>    0006      0      0      0      0      0      0      0      0      0      0
#>    0009      0      0      0      0      0      0      0      0      0      0
#>    0020      0      0      0      0      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      0
#>    0100      0      0      0      0      0      0      0      0      0      0
#>    0101      0      0      0      0      0      0      0      0      0      0
#>    0102      0      0      0      0      0      0      0      0      0      0
#>        sp
#> quadrat psycde psycho psycma pterro quaras quasam randar rinosy sapiau sennda
#>    0000      0      0      0      0      0      0      0      0      0      0
#>    0002      0      0      0      0      0      0      0      0      0      0
#>    0006      0      0      0      0      0      0      0      0      0      0
#>    0009      0      0      0      0      0      0      0      0      0      0
#>    0020      0      0      0      0      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      0
#>    0100      0      1      0      0      0      0      0      0      0      0
#>    0101      0      0      0      0      0      0      0      0      0      0
#>    0102      0      0      0      0      0      0      0      0      0      0
#>        sp
#> quadrat simaam sipagu sloate socrex soroaf sponmo stylst swars1 swars2 sympgl
#>    0000      0      0      0      0      1      0      0      0      0      0
#>    0002      0      0      0      0      0      0      0      0      0      0
#>    0006      0      0      0      0      0      0      0      0      0      0
#>    0009      0      0      0      0      0      0      0      0      0      0
#>    0020      0      0      0      0      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      1
#>    0100      0      0      0      0      0      0      0      0      0      0
#>    0101      0      0      0      0      0      0      0      0      0      0
#>    0102      0      0      0      0      0      0      0      0      0      0
#>        sp
#> quadrat tab2ar tachve talipr tet2pa tri2pa tri2tu tropra turpoc virose vochfe
#>    0000      0      0      0      0      0      0      0      0      0      0
#>    0002      0      0      0      0      0      0      0      0      0      0
#>    0006      0      0      0      0      0      0      0      0      0      0
#>    0009      0      0      0      0      0      0      0      0      0      0
#>    0020      0      0      0      1      0      0      0      0      0      0
#>    0022      0      0      0      0      0      0      0      0      0      0
#>    0100      0      0      0      0      0      0      0      0      0      0
#>    0101      0      0      0      0      0      0      0      0      0      0
#>    0102      0      0      0      0      0      0      0      0      0      0
#>        sp
#> quadrat xyl1ma
#>    0000      0
#>    0002      0
#>    0006      0
#>    0009      0
#>    0020      0
#>    0022      0
#>    0100      0
#>    0101      0
#>    0102      0
#>  [ reached getOption("max.print") -- omitted 434 rows ]

class(n)

#> [1] "table"

abundance_n(stem, group_by = "sp", FALSE)

#> 
#> acaldi aegipa alchco alibed allops alsebl anaxpa andiin annoac annosp apeime 
#>     16      1      1      3      1     27      3      2      4      1      2 
#> aspicr bactc1 bactc2 beilpe brosal calolo cappfr caseac casear casesy cassel 
#>      2      3      1      8      3      6      9      3      1      1      8 
#> cecrin cecrob cedrod celtsc cha2sc chr1ec chr2ar coccco coccma cordal cordbi 
#>      7      2      1      1      1      4      2      1      3      2      6 
#> cordla cou2cu crotbi cupasy dendar des2pa dio2ar drypst ery2ma eugeco eugega 
#>      9      9      4      4      1     60      1      6      2      5      3 
#> eugene eugeoe faraoc gar2in guapst guargu guarsp guatdu guetfo gustsu hameax 
#>      1     12     86     10      2      9      3      4      1      2      1 
#> hampap hassfl heisco herrpu hirttr hybapr ingafa ingago ingama ingami ingape 
#>      1      5      1      1     10    169      1      5      5      2      1 
#> ingapu ingaqu ingas1 ingasa ingath ingaum laciag laetth licapl loncla luehse 
#>      2      3      3      1      1      7     11      2      1      2      3 
#> malmsp maquco micoaf micoar micone mourmy nectgl nects1 ocotce ocotob ocotpu 
#>      3      4      1      2      3     29      2      1      1      3      2 
#> ocotwh oenoma olmeas ormocr ouralu paligu pentma picrla pipeae pipeco pla1pi 
#>      2      3      2      1      9      9      2      5      1     13      1 
#> poular poutre pri2co protco protpa protte pse1se psycac psycde psycg3 psycho 
#>      9      7      6      4      6     10      1      1      1      1     30 
#> psycma pterro quaras quasam randar rinosy sapiau sennda simaam sipagu sipapa 
#>      4      9      8      1      5     15      1      2     15      1      1 
#> sloate socrex soroaf sponmo stylst swars1 swars2 sympgl tab2ar tachve talipr 
#>      1      8     10      1      2     10      3      1      6     13      3 
#> tet2pa tri2pa tri2tu tropra turpoc unonpi urerba virose virosu vochfe xyl1ma 
#>     18      4     49      3      2      2      1      9      1      1      6 
#> zuelgu 
#>      1