Basal area of trees.

ba(): calculates the individual basal areas in square meters for all dbh-values submitted.

ba_sum(): calculates the basal area summed over all submitted dbh values (after removing NAs).

ba(dbh, dbhunit = "mm")

ba_sum(dbh, dbhunit = "mm", mindbh = 10)

Arguments

dbh
Stem diameter, usually at breast height (see data dictionary at http://ctfs.si.edu/Public/DataDict/data_dict.php).

dbhunit
'cm' or 'mm', only used for basal area.

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

Value

ba(): A vector of basal area values of same length as the submitted vector of dbhs.

ba_sum(): A number giving the the basal area summed over all submitted dbhs.

Examples

# Takes NAs and throws informative warning
(dbh <- c(seq(5, 50, by = 5), NA))

#>  [1]  5 10 15 20 25 30 35 40 45 50 NA

ba(dbh, dbhunit = "cm")  # default dbhunit is milimeters ("mm")

#> Warning: is_not_na : The values of dbh are sometimes NA.
#> There was 1 failure:
#>   Position Value   Cause
#> 1       11    NA missing

#>  [1] 0.001963495 0.007853982 0.017671459 0.031415927 0.049087385 0.070685835
#>  [7] 0.096211275 0.125663706 0.159043128 0.196349541          NA

# (From now on, suppressing warnings to avoid distraction)

# Filter stems at or over a minimum dbh and sum them, all in one step
suppressWarnings(
  ba_sum(dbh, mindbh = 30)
)

#> [1] 0.006479535

# Same; a bit longer but may be easier to understand what is going on
over_mindbh <- dbh[dbh >= 30]
suppressWarnings(
  sum(ba(over_mindbh), na.rm = TRUE)
)

#> [1] 0.006479535