Create tables of abundance and basal area by (round mean) year.

Create tables of abundance and basal area by (round mean) year.

byyr_abundance(vft)

byyr_basal_area(vft)

Arguments

vft
A dataframe; particularly a ForestGEO ViewFullTable.

Value

A dataframe.

Examples

library(fgeo.tool)
library(fgeo.base)

vft <- data.frame(
  Tag = c("0001", "0001", "0002", "0002"),
  PlotName = "p",
  Status = c("dead", "dead", "alive", "alive"),
  DBH = c(NA, NA, 10, 100),
  ExactDate = c("2000-01-01", "2001-01-01", "2000-01-01", "2001-01-01"),
  PlotCensusNumber = c(1, 2, 1, 2),
  CensusID = c(1, 2, 1, 2),
  Genus = c("A", "A", "B", "B"),
  SpeciesName = c("a", "a", "b", "b"),
  Family = "f",
  stringsAsFactors = FALSE
)
vft

#>    Tag PlotName Status DBH  ExactDate PlotCensusNumber CensusID Genus
#> 1 0001        p   dead  NA 2000-01-01                1        1     A
#> 2 0001        p   dead  NA 2001-01-01                2        2     A
#> 3 0002        p  alive  10 2000-01-01                1        1     B
#> 4 0002        p  alive 100 2001-01-01                2        2     B
#>   SpeciesName Family
#> 1           a      f
#> 2           a      f
#> 3           b      f
#> 4           b      f

# First pick the data you want
pick1 <- fgeo.base::pick_plotname(vft, "p")

#> Warning: restarting interrupted promise evaluation

#> Warning: internal error -3 in R_decompress1

#> Error in get0(oNam, envir = ns): lazy-load database 'C:/Users/LeporeM/Documents/R/R-3.5.0/library/fgeo.base/R/fgeo.base.rdb' is corrupt

pick2 <- drop_dead_trees_by_cns(pick1)

#> Warning: internal error -3 in R_decompress1

#> Error in get0(oNam, envir = ns): lazy-load database 'C:/Users/LeporeM/Documents/R/R-3.5.0/library/fgeo.base/R/fgeo.base.rdb' is corrupt

pick3 <- pick_dbh_min(pick2, 10)

#> Warning: internal error -3 in R_decompress1

#> Error in eval(expr, envir, enclos): lazy-load database 'C:/Users/LeporeM/Documents/R/R-3.5.0/library/fgeo.base/R/fgeo.base.rdb' is corrupt

pick3

#> Error in eval(expr, envir, enclos): object 'pick3' not found

byyr_abundance(pick3)

#> Error in is.data.frame(vft): object 'pick3' not found

ba <- byyr_basal_area(pick3)

#> Error in is.data.frame(vft): object 'pick3' not found

ba

#> Error in eval(expr, envir, enclos): object 'ba' not found

# Convert units and standardize by plot size in hectares
years <- c("2000", "2001")
in_he <- conv_unit_at(ba, .at = years, from = "mm2", to = "hectare")

#> Error in inv_which(.x, .at): object 'ba' not found

standardize_at(in_he, .at = years, total = 50)

#> Error in inv_which(.x, .at): object 'in_he' not found