Count distinct values of treeid, stemid, or any variable by groups.

Opinions vary on what abundance means. For example, to calculate the abundance of woods with ForestGEO data, you may need to define the status and dbh range of the stems you are interested. But that's not the job of this functions. These functions only count distinct occurrences of variables. Thus, to calculate abundance in the exact way you want you need to compose these functions (e.g. collapse_treeid_max(), pick_dbh_min(), pick_status(), drop_dead_tree()). To avoid creating false expectations, the name of these functions intentionally avoid the word abundance.

These functions are usually used with grouped data (via dplyr::group_by()). The output will have one row for each group.

count_distinct() counts distinct occurrences of any variable of any dataset (not specifically a ForestGEO dataset).

count_distinct_stemid() and count_distinct_treeid() are simpler and specialized versions count_distinct(), for specifically counting distinct occurrences of the variables TreeID and StemID, or stemID and treeID, which uniquely identify each stem and tree in ViewFullTables and census datasets of ForestGEO.

count_distinct(.data, .var)

count_distinct_stemid(.data)

count_distinct_treeid(.data)

Arguments

.data
A dataframe, commonly grouped with group_by().

.var
A variable to count distinct occurrences.

Value

An object of the same class as .data. One grouping level will be dropped.

Warning

These functions do not remove dead stems or trees. If you don't want dead trees to be included, remove them first (see drop_dead_tree(), drop_dead_stem().

See also

Examples

census <- tibble::tibble(
  treeID = c(1, 1, 2, 3, 3, 3),
  stemID = c(1, 2, 3, 4, 5, 6),
  quadrat = paste0("000", rep(1:2, each = 3)),
  sp = c(paste0("sp", c(1, 1, 2)), paste0("sp", c(3, 3, 3))),
  dbh = abs(sample(rnorm(100), 6) * 10)
)
census

#> # A tibble: 6 x 5
#>   treeID stemID quadrat sp      dbh
#>    <dbl>  <dbl> <chr>   <chr> <dbl>
#> 1      1      1 0001    sp1   18.8 
#> 2      1      2 0001    sp1    1.08
#> 3      2      3 0001    sp2    7.96
#> 4      3      4 0002    sp3    8.95
#> 5      3      5 0002    sp3    5.58
#> 6      3      6 0002    sp3   15.3 

count_distinct(census, stemID)

#> # A tibble: 1 x 1
#>       n
#>   <int>
#> 1     6

by_quad <- group_by(census, quadrat)
count_distinct(by_quad, stemID)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        3
#> 2 0002        3

count_distinct(by_quad, treeID)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        2
#> 2 0002        1

count_distinct_stemid(by_quad)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        3
#> 2 0002        3

count_distinct_treeid(by_quad)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        2
#> 2 0002        1

by_sp <- group_by(census, sp)
count_distinct_stemid(by_sp)

#> # A tibble: 3 x 2
#>   sp        n
#>   <chr> <int>
#> 1 sp1       2
#> 2 sp2       1
#> 3 sp3       3

count_distinct_treeid(by_sp)

#> # A tibble: 3 x 2
#>   sp        n
#>   <chr> <int>
#> 1 sp1       1
#> 2 sp2       1
#> 3 sp3       1

by_quad_sp <- group_by(census, quadrat, sp)
count_distinct_stemid(by_quad_sp)

#> # A tibble: 3 x 3
#> # Groups:   quadrat [?]
#>   quadrat sp        n
#>   <chr>   <chr> <int>
#> 1 0001    sp1       2
#> 2 0001    sp2       1
#> 3 0002    sp3       3

count_distinct_treeid(by_quad_sp)

#> # A tibble: 3 x 3
#> # Groups:   quadrat [?]
#>   quadrat sp        n
#>   <chr>   <chr> <int>
#> 1 0001    sp1       1
#> 2 0001    sp2       1
#> 3 0002    sp3       1

# Compare to dplyr::summarize() and dplyr::n_distinct()
library(dplyr)

by_cyl <- group_by(mtcars, cyl)
count_distinct(by_cyl, mpg)

#> # A tibble: 3 x 2
#>     cyl     n
#>   <dbl> <int>
#> 1     4     9
#> 2     6     6
#> 3     8    12

# Same
summarize(by_cyl, n = n_distinct(mpg))

#> # A tibble: 3 x 2
#>     cyl     n
#>   <dbl> <int>
#> 1     4     9
#> 2     6     6
#> 3     8    12

n_distinct(mtcars)

#> [1] 32

# Same
nrow(mtcars)

#> [1] 32

# NOT RUN {
# Fails; you must supply `.var` to count distinct values if that variable
count_distinct(mtcars)
# }

count_distinct(mtcars, cyl)

#>   n
#> 1 3

# Same
n_distinct(mtcars$cyl)

#> [1] 3

# Same
length(unique(mtcars$cyl))

#> [1] 3

# NOT RUN {
# Fails because it expect not column names but additional vectors (via `...`)
dplyr::n_distinct(mtcars, cyl)
# }