Count distinct values of treeid, stemid (optionally by groups).

These functions are simpler, specialized versions of count_distinct(), specifically for counting distinct occurrences of the variables stemID and treeID (or TreeID and StemID) which uniquely identify each stem and tree in ForestGEO-like datasets. They work with data grouped with dplyr::group_by().

count_distinct_treeid(.data)

count_distinct_stemid(.data)

Arguments

.data
A ForestGEO-like dataframe, commonly grouped with group_by().

Value

An object of the same class as .data. One grouping level will be dropped (because of dplyr::summarize()).

Details

count_unique_treeid() throws an error if each data-group contains more than one value of treeid (i.e. if it contains multiple stems). You should first collapse treeid by picking a single stem per treeid per group. Both count_unique_treeid() and count_unique_stemid() warn if the dataset contains multiple censusid.

Opinions vary on what abundance means. For example, to calculate the abundance of woods with ForestGEO data, you may need to define the status and dbh range of the stems you are interested. But that's not the job of this functions. These functions only count distinct occurrences of variables.

Warning

These functions do not remove dead stems or trees. If you don't want dead trees to be included, remove them first (see drop_dead_tree(), drop_dead_stem().

See also

Examples

library(dplyr)

census <- tibble::tibble(
  treeID = c(1, 1, 2, 3, 3, 3),
  stemID = c(1, 2, 3, 4, 5, 6),
  quadrat = paste0("000", rep(1:2, each = 3)),
  sp = c(paste0("sp", c(1, 1, 2)), paste0("sp", c(3, 3, 3))),
  dbh = abs(sample(rnorm(100), 6) * 10)
)
census

#> # A tibble: 6 x 5
#>   treeID stemID quadrat sp      dbh
#>    <dbl>  <dbl> <chr>   <chr> <dbl>
#> 1      1      1 0001    sp1   18.8 
#> 2      1      2 0001    sp1    1.08
#> 3      2      3 0001    sp2    7.96
#> 4      3      4 0002    sp3    8.95
#> 5      3      5 0002    sp3    5.58
#> 6      3      6 0002    sp3   15.3 

by_quad <- group_by(census, quadrat)
count_distinct_stemid(by_quad)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        3
#> 2 0002        3

count_distinct_treeid(by_quad)

#> Error:   Detected duplicated values of treeid.
#>     Expected unique values of treeid within each data-group (if any)

by_sp <- group_by(census, sp)
count_distinct_stemid(by_sp)

#> # A tibble: 3 x 2
#>   sp        n
#>   <chr> <int>
#> 1 sp1       2
#> 2 sp2       1
#> 3 sp3       3

count_distinct_treeid(by_sp)

#> Error:   Detected duplicated values of treeid.
#>     Expected unique values of treeid within each data-group (if any)

by_quad_sp <- group_by(census, quadrat, sp)
count_distinct_stemid(by_quad_sp)

#> # A tibble: 3 x 3
#> # Groups:   quadrat [2]
#>   quadrat sp        n
#>   <chr>   <chr> <int>
#> 1 0001    sp1       2
#> 2 0001    sp2       1
#> 3 0002    sp3       3

count_distinct_treeid(by_quad_sp)

#> Error:   Detected duplicated values of treeid.
#>     Expected unique values of treeid within each data-group (if any)