Abundance, richness, singletons: Count individuals in total or by groups.

count(), tally(), add_count() and add_tally() help you count rows in a dataframe. They come from the dplyr package and are reexported by fgeo.abundance for your convenience (see reexports). For details see dplyr::count().

count_duplicated helps you count duplicated observations of a variable. It is a shortcut for dplyr::filter(dplyr::count(x, ...) > 1)).

count_duplicated(x, ..., wt = NULL, sort = FALSE)

Arguments

x
A dataframe.

...
Variables to group by.

wt
(Optional) If omitted (and no variable named n exists in the data), will count the number of rows. If specified, will perform a "weighted" tally by summing the (non-missing) values of variable wt. A column named n (but not nn or nnn) will be used as weighting variable by default in tally(), but not in count(). This argument is automatically quoted and later evaluated in the context of the data frame. It supports unquoting. See vignette("programming") for an introduction to these concepts.

sort
if TRUE will sort output in descending order of n

Value

A dataframe.

Details

The issue of calculating the abundance of individuals in a sample is a particular case of counting observations in general. By counting individuals you can also calculate richness, find singletons, or duplicated observations of a variable. The examples below show how the most commonly used functions may be applied to ForestGEO data. To learn more about these functions, and to see more general examples see ?dplyr::count().

All of these functions quote some argument to make interactive use easier. If you want to use them inside your own functions you should learn about tidy eval (implemented via the rlang package). A good place to start is at dplyr's website.

See also

Examples

census <- tibble::tibble(
  stemID = 1:6,
  quadrat = paste0("000", rep(1:2, each = 3)),
  sp = c(paste0("sp", c(1, 1, 2)), paste0("sp", c(3, 3, 3))),
  dbh = abs(sample(rnorm(100), 6) * 10)
)
census

#> # A tibble: 6 x 4
#>   stemID quadrat sp      dbh
#>    <int> <chr>   <chr> <dbl>
#> 1      1 0001    sp1   18.8 
#> 2      2 0001    sp1    1.08
#> 3      3 0001    sp2    7.96
#> 4      4 0002    sp3    8.95
#> 5      5 0002    sp3    5.58
#> 6      6 0002    sp3   15.3 

# Abundance (rows count) by quadrat
count(census, quadrat)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        3
#> 2 0002        3

# Abundance (rows count) by quadrat by species
count(census, quadrat, sp)

#> # A tibble: 3 x 3
#>   quadrat sp        n
#>   <chr>   <chr> <int>
#> 1 0001    sp1       2
#> 2 0001    sp2       1
#> 3 0002    sp3       3

# Richness by quadrat:
# Count is designed so that you can call it repeatedly, each time rolling up
# a level of detail.
# Now, each row by quadrat is a unique species, so counting rows gives
# richness
n_quad_sp <- count(census, quadrat, sp)
richness_quad <- count(n_quad_sp, quadrat)
richness_quad

#> # A tibble: 2 x 2
#>   quadrat    nn
#>   <chr>   <int>
#> 1 0001        2
#> 2 0002        1

# Singleton:
# add_count() is useful for groupwise filtering e.g.: show only species that
# have a single member
out <- add_count(census, quadrat, sp)
subset(out, n == 1)

#> # A tibble: 1 x 5
#>   stemID quadrat sp      dbh     n
#>    <int> <chr>   <chr> <dbl> <int>
#> 1      3 0001    sp2    7.96     1

# Or the complement
subset(out, n > 1)

#> # A tibble: 5 x 5
#>   stemID quadrat sp      dbh     n
#>    <int> <chr>   <chr> <dbl> <int>
#> 1      1 0001    sp1   18.8      2
#> 2      2 0001    sp1    1.08     2
#> 3      4 0002    sp3    8.95     3
#> 4      5 0002    sp3    5.58     3
#> 5      6 0002    sp3   15.3      3

# A shortcut to find duplicates that drops irrelevant columns:
count_duplicated(census, quadrat, sp)

#> # A tibble: 2 x 3
#>   quadrat sp        n
#>   <chr>   <chr> <int>
#> 1 0001    sp1       2
#> 2 0002    sp3       3

# Any duplicated stemID?
count_duplicated(census, stemID)

#> # A tibble: 0 x 2
#> # ... with 2 variables: stemID <int>, n <int>