Number of individuals in total or within groups.

In an ecological context, these functions calculate abundance. To choose the status of the individuals you want to count, use count_status(). If you only want to count alive individuals, you may use the shortcut count_alive().

count_status(x, .status, ...)

count_alive(x, ...)

Arguments

x
Census.

.status
A character vector giving the statuses to keep.

...
Optional. Bare name of variables to group by (passed to dplyr::count()).

Value

If you provide grouping variables, these functions output a grouped summary. Else, these functions output a tally.

See also

dplyr::count().

Examples

library(bciex)

stems <- bci12s5mini
trees <- bci12t5mini

# Tally of alive individuals
count_alive(stems)

#> [1] 668

# Count individuals of any status, say, dead ones
count_status(trees, .status = "D")

#> [1] 284

# Give one or multiple grouping variables to get a grouped summary
# Alive
count_alive(stems, sp)

#> # A tibble: 144 x 2
#>       sp     n
#>    <chr> <dbl>
#> 1 apeime     2
#> 2 alchco     1
#> 3 quaras     6
#> 4 beilpe     5
#> 5 luehse     6
#> 6 casear     1
#> # ... with 138 more rows

# Dead
count_status(stems, .status = "D", sp)

#> # A tibble: 144 x 2
#>       sp     n
#>    <chr> <dbl>
#> 1 apeime     0
#> 2 alchco     0
#> 3 quaras     2
#> 4 beilpe     2
#> 5 luehse     1
#> 6 casear     0
#> # ... with 138 more rows

# Multiple groups
# Alive
stems2 <- rbind(bci12s5mini, bci12s6mini)
count_alive(stems2, sp, CensusID)

#> # A tibble: 244 x 3
#>       sp CensusID     n
#>    <chr>    <int> <dbl>
#> 1 apeime        5     2
#> 2 apeime        6     2
#> 3 alchco        5     1
#> 4 alchco        6     2
#> 5 quaras        5     6
#> 6 quaras        6     7
#> # ... with 238 more rows

# Dead
stems2 <- rbind(bci12s5mini, bci12s6mini)
count_status(stems2, .status = "D", sp, CensusID)

#> # A tibble: 219 x 3
#>       sp CensusID     n
#>    <chr>    <int> <dbl>
#> 1 apeime       NA     0
#> 2 alchco       NA     0
#> 3 quaras        5     1
#> 4 quaras       NA     3
#> 5 beilpe        5     2
#> 6 beilpe        6     1
#> # ... with 213 more rows