Count unique trees based on the dbh of their largest stem.

This functions count unique treeid by censusid (automatically) and any number of additional grouping variables (via dplyr::group_by(). Data with multi-stem trees is first collapsed. The collapse defaults to picking the stem (of each treeid) with maximum dbh (pick minimum dbh with the argument .collapse).

count_woods() is general and allows filtering data using any number of expressions passed via the argument ... to dplyr::filter().

count_trees() is an opinionated shortcut that defines tress as stems of dbh of 100 mm and above.

count_samplings() is an opinionated shortcut that defines saplings as stems of dbh of between 10 mm inclusive and 100 mm exclusive.

count_woods(.data, ..., .collapse = collapse_treeid_max)

count_trees(.data, .collapse = collapse_treeid_max)

count_saplings(.data, .collapse = collapse_treeid_max)

Arguments

.data
A dataframe; particularly a ForestGEO census or ViewFullTable.

...
Expressions to pick stems of specific dbh -- where stems refers to the largest stem of each tree.

.collapse
collapse_treeid_max and collapse_treeid_min(bare name of the corresponding functions) collapse the data picking the stem (of each treeid) with the maximum and minimum dbh, respectively.

Value

A dataframe with preserved groups (if any).

Examples

library(dplyr)

census <- tibble::tribble(
  ~dbh,   ~sp, ~treeID, ~stemID,
    10, "sp1",     "1",   "1.1",
   100, "sp1",     "1",   "1.2",
    22, "sp2",     "2",   "2.1",
    99, "sp2",     "2",   "2.2",
    NA, "sp2",     "2",   "2.3"
)

# Count all woods
count_woods(census)

#> Error in multiple_plotname(.x): could not find function "multiple_plotname"

by_sp_treeid <- group_by(census, sp, treeID)
count_trees(by_sp_treeid)

#> Error in multiple_plotname(.x): could not find function "multiple_plotname"

by_sp <- group_by(census, sp)
count_woods(by_sp)

#> Error in multiple_plotname(.x): could not find function "multiple_plotname"

count_woods(by_sp, dbh >= 100)

#> Error in multiple_plotname(.x): could not find function "multiple_plotname"

# Same
count_trees(by_sp)

#> Error in multiple_plotname(.x): could not find function "multiple_plotname"

count_woods(by_sp, dbh > 10, dbh < 100)

#> Error in multiple_plotname(.x): could not find function "multiple_plotname"

# Same
count_saplings(by_sp)

#> Error in multiple_plotname(.x): could not find function "multiple_plotname"