These functions are alternatives which source code use dplyr or plyr.

Compared to using base exclusively, these functions may be a little more friendly for users, but maybe harder for developers to understand and maintain. dplyr evolves very rapidly and therefore some functions that are recommended today may be deprecated later. Yet, the source code is available so it is always possible to re-use it into forestr.

n_dplyr(x, ..., only_alive = TRUE)

n_dplyr_se(x, groups = c("status", "sp"), only_alive = TRUE)

n_plyr(x, groups = c("sp", "status"), only_alive = TRUE, ...)

Arguments

x
Census.

...
Bare names of variables to group by.

only_alive
If TRUE, filter only alive individuals.

groups
Character string giving the names of variables to group by.

Value

Tally if groups = NULL, else, a count by group.

Details

Grouping with a character string

n_dplyr_se() uses latest tidy eval approach.

n_plyr() uses plyr (deprecated).

Grouping with bare variable names:

n_dplyr() uses the latest "tidy eval" approach.

Examples

library(dplyr)

stem <- bci12s7mini

# Giving grouping variables in a character string  -----------------------

# USING DPLYR (IF NEW TO DPLYR, THE SOURCE CODE MAY BE HARDER TO UNDERSTAND)

# Standard evaluation using the newest (tidy eval) approach.

n_dplyr_se(stem)

#> # A tibble: 111 x 3
#>   status     sp     n
#>    <chr>  <chr> <int>
#> 1      A acaldi     5
#> 2      A alibed     2
#> 3      A alsebl    21
#> 4      A anaxpa     3
#> 5      A andiin     1
#> 6      A annoac     5
#> # ... with 105 more rows

# Same:
n_dplyr_se(stem, groups = c("status", "sp"))

#> # A tibble: 111 x 3
#>   status     sp     n
#>    <chr>  <chr> <int>
#> 1      A acaldi     5
#> 2      A alibed     2
#> 3      A alsebl    21
#> 4      A anaxpa     3
#> 5      A andiin     1
#> 6      A annoac     5
#> # ... with 105 more rows

# NULL groups result in a tally.
n_dplyr_se(stem, groups = NULL)

#>     n
#> 1 665

# Count not only alive
n <- n_dplyr_se(stem, groups = c("status", "sp"), only_alive = FALSE)
arrange(n, sp)

#> # A tibble: 307 x 3
#>   status     sp     n
#>    <chr>  <chr> <int>
#> 1      A acaldi     5
#> 2      D acaldi    11
#> 3      G acaldi    11
#> 4      D aegipa     1
#> 5      D alchco     1
#> 6      G alchco    10
#> # ... with 301 more rows

# INSTEAD OF DPLYR, USING PLYR (DEPRECATED, AND SLOWER THAN BASE AND DPLYR)

# Using plyr, the deprecated ancestor of dplyr

head(n_plyr(stem))

#>        sp status  n
#> 1  acaldi      A  5
#> 7  alibed      A  2
#> 10 alsebl      A 21
#> 13 anaxpa      A  3
#> 16 andiin      A  1
#> 19 annoac      A  5

head(n_plyr(stem), groups = c("quadrat", "sp"), only_alive = FALSE)

#>        sp status  n
#> 1  acaldi      A  5
#> 7  alibed      A  2
#> 10 alsebl      A 21
#> 13 anaxpa      A  3
#> 16 andiin      A  1
#> 19 annoac      A  5

df <- data.frame(
  id = 1:6,
  sp = letters[c(1, 2, 2, 3, 3, 3)],
  status = rep(c("A", "D"), each = 3),
  stringsAsFactors = FALSE
)
df

#>   id sp status
#> 1  1  a      A
#> 2  2  b      A
#> 3  3  b      A
#> 4  4  c      D
#> 5  5  c      D
#> 6  6  c      D

# The same ----

dplyr::count(df, sp, status)

#> # A tibble: 3 x 3
#>      sp status     n
#>   <chr>  <chr> <int>
#> 1     a      A     1
#> 2     b      A     2
#> 3     c      D     3

grouped <- dplyr::group_by(df, sp, status)
dplyr::summarise(grouped, n = n())

#> # A tibble: 3 x 3
#> # Groups:   sp [?]
#>      sp status     n
#>   <chr>  <chr> <int>
#> 1     a      A     1
#> 2     b      A     2
#> 3     c      D     3

tibble::as_tibble(
  n_plyr(df, c("sp", "status"), only_alive = FALSE)
)

#> # A tibble: 3 x 3
#>      sp status     n
#>   <chr>  <chr> <int>
#> 1     a      A     1
#> 2     b      A     2
#> 3     c      D     3

# Giving grouping variables as bare names --------------------------------

# dplyr defaults to using bare variable names. 

n_dplyr(stem)

#> # A tibble: 1 x 1
#>       n
#>   <int>
#> 1   665

n_dplyr(stem, sp, only_alive = FALSE)

#> # A tibble: 144 x 2
#>       sp     n
#>    <chr> <int>
#> 1 acaldi    27
#> 2 aegipa     1
#> 3 alchco    11
#> 4 alibed     3
#> 5 allops     1
#> 6 alsebl    36
#> # ... with 138 more rows