Count and basal area of neighboring stems.

Includes all values of status and dbh. You should pick the values you want before using these functions (e.g. to drop dead stems).

count_neighbor(.data, .subset = NULL, r, plotdim = NULL)

basal_area_neighbor(.data, .subset = NULL, r, plotdim = NULL)

Arguments

.data
A Dataframe; particularly a ForestGEO census table (tree or stem).

.subset
An optional dataframe giving the (gx, gy) coordinates of specific individuals (tag and sp) for which to count neighbors. For example, .subset may be a subset of only one species; or seedling that are not part of census .data. .subset must have columns gx, gy, sp, and tag.

r
Radius.

plotdim
The x and y dimensions of the plot.

Value

A dataframe.

A dataframe with two columns conspecific and heterospecific, and optionally one extra column for each variable used to group by. The number of rows is as follows:

If .subset is NULL, the output and .data have equal number of rows, both for grouped and ungrouped .data.

If .subset is not NULL, and .data is ungrouped, the output and .subset have equal number of rows.

If .subset is not NULL, and .data is grouped, the output has a number of rows that equals that of .subset multiplied by the number of groups.

Warning

This function has a pending issue by which a dataset with a single row does not result in zero conspecific neighbors (https://github.com/forestgeo/fgeo.abundance/issues/68).

Examples

tree <- fgeo.data::luquillo_tree5_random
# Guess `plotdim`
count_neighbor(tree, r = 20)

#> Gessing: plotdim = c(320, 500)

#> Total elapsed time ~ 0.87 seconds 

#> # A tibble: 1,004 x 2
#>    conspecific heterospecific
#>          <dbl>          <dbl>
#>  1        1.23           6.15
#>  2        2              6   
#>  3        2              7   
#>  4        1              9   
#>  5        1              1   
#>  6        1             11   
#>  7        1              5   
#>  8        2              6   
#>  9        1              3   
#> 10        2              7   
#> # ... with 994 more rows

count_neighbor(tree, r = 20, plotdim = c(320, 500))

#> Total elapsed time ~ 0.53 seconds 

#> # A tibble: 1,004 x 2
#>    conspecific heterospecific
#>          <dbl>          <dbl>
#>  1        1.23           6.15
#>  2        2              6   
#>  3        2              7   
#>  4        1              9   
#>  5        1              1   
#>  6        1             11   
#>  7        1              5   
#>  8        2              6   
#>  9        1              3   
#> 10        2              7   
#> # ... with 994 more rows

basal_area_neighbor(tree, r = 20, plotdim = c(320, 500))

#> Total elapsed time ~ 0.55 seconds 

#> # A tibble: 1,004 x 2
#>    conspecific heterospecific
#>          <dbl>          <dbl>
#>  1      360.             614.
#>  2       14.3            602.
#>  3       17.8            485.
#>  4        8.55          1258.
#>  5      154.             172.
#>  6      475.             726.
#>  7      243.            3180.
#>  8      455.             200.
#>  9     2865.             412.
#> 10       68.5            386.
#> # ... with 994 more rows

# Notice how the number of rows of the output varies with the input -------

tree <- tibble::tribble(
  ~gx, ~gy, ~tag,   ~sp,  ~dbh, ~status,
  5,   5, "01", "sp1",     5,     "A",
  5,   5, "02", "sp1",     5,     "A",
  5,   5, "03", "sp2",     5,     "A",
  5,   5, "04", "sp2",     5,     "A"
)

# `.subset = NULL`, ungrouped `.data`
count_neighbor(.data = tree, .subset = NULL, r = 20, plotdim = c(320, 500))

#> Total elapsed time ~ 0 seconds 

#> # A tibble: 4 x 2
#>   conspecific heterospecific
#>         <dbl>          <dbl>
#> 1        4.68           4.68
#> 2        4.68           4.68
#> 3        4.68           4.68
#> 4        4.68           4.68

# `.subset = NULL`, grouped `.data`
count_neighbor(tree, .subset = NULL, r = 20, plotdim = c(320, 500))

#> Total elapsed time ~ 0 seconds 

#> # A tibble: 4 x 2
#>   conspecific heterospecific
#>         <dbl>          <dbl>
#> 1        4.68           4.68
#> 2        4.68           4.68
#> 3        4.68           4.68
#> 4        4.68           4.68

by_sp <- dplyr::group_by(tree, sp)
count_neighbor(.data = by_sp, .subset = NULL, r = 20, plotdim = c(320, 500))

#> Total elapsed time ~ 0 seconds 

#> Total elapsed time ~ 0 seconds 

#> # A tibble: 4 x 3
#>   sp    conspecific heterospecific
#>   <chr>       <dbl>          <dbl>
#> 1 sp1          4.68              0
#> 2 sp1          4.68              0
#> 3 sp2          4.68              0
#> 4 sp2          4.68              0

subset <- tibble::tribble(
  ~gx, ~gy, ~tag,   ~sp,
  3,   3, "99", "sp1"
)

# `.subset` not `NULL`, ungrouped `.data`
count_neighbor(.data = tree, .subset = subset, r = 20, plotdim = c(320, 500))

#> Total elapsed time ~ 0 seconds 

#> # A tibble: 1 x 2
#>   conspecific heterospecific
#>         <dbl>          <dbl>
#> 1        5.68           5.68

# `.subset` not `NULL`, ungrouped `.data`
by_sp <- dplyr::group_by(tree, sp)
count_neighbor(.data = by_sp, .subset = subset, r = 20, plotdim = c(320, 500))

#> Total elapsed time ~ 0 seconds 

#> Total elapsed time ~ 0 seconds 

#> # A tibble: 2 x 3
#>   sp    conspecific heterospecific
#>   <chr>       <dbl>          <dbl>
#> 1 sp1          5.68           0   
#> 2 sp2          0              5.68