Recruitment and mortality

Caculate demography metrics accross the entire dataset or within one or two groups (as in (http://ctfs.si.edu/Public/CTFSRPackage/). Each of these demography functions has a *_df() version that accepts not more than one split variable and, instead of a list, returns a dataframe.

For every individual of all species you must provide two complete datasets, one per census, with variables dbh, pom, status and date. Any status indicating a live tree can be submitted in the variable alivecode.

mortality(census1, census2, alivecode = c("A", "AB", "AS"), split1 = NULL,
  split2 = NULL)

mortality_df(census1, census2, alivecode = c("A", "AB", "AS"),
  split1 = NULL)

recruitment(census1, census2, mindbh = 10, alivecode = c("A", "AB", "AS"),
  split1 = NULL, split2 = NULL)

recruitment_df(census1, census2, mindbh = 10, alivecode = c("A", "AB",
  "AS"), split1 = NULL)

Arguments

census1
The R Analytical Table for a single census, either tree or stem.

census2
The matching R Analytical Table for a later census.

alivecode
character, codes of the variable `status` that indicate the tree is alive. The default 'A' is the standard CTFS designation for living trees or stems

split1
a vector of categories, one per individual

split2
another vector of categories, one per individual

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

Value

(xxx ask documentation) recruitment() returns a list with components:

N2:

R:

rate:

lower:

upper:

time:

date1:

date2:

mortality() returns a list with components:

N: the number of individuals alive in the census 1 per category selected

D: the number of individuals no longer alive in census 2

rate: the mean annualized mortality rate constant per category selected, calculated as (log(N)-log(S))/time

upper: upper confidence limit of mean rate

lower: lower confidence limit of mean rate

time: mean time interval in years

date1: mean date included individuals were measured in census 1, as julian object (R displays as date, but treats as integer)

date2: mean date in census 2

dbhmean: mean dbh in census 1 of individuals included

Details

Survivors are all individuals alive in both censuses, with status == A in the first census, and larger than the minimum dbh in the first census. The total population in the second census includes all those alive, above the minimum dbh, plus any other survivors. Individuals whose status is NA in either census are deleted from all calculations.

See also

Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: growth

Other functions to compare two censuses: growth

Examples

# Some data to play with
census1 <- forestr::bci12t6mini
census2 <- forestr::bci12t7mini

recruitment(census1, census2)

#> $N2
#> [1] 554
#> 
#> $R
#> [1] 71
#> 
#> $rate
#> [1] 0.0276
#> 
#> $lower
#> [1] 0.0219
#> 
#> $upper
#> [1] 0.0348
#> 
#> $time
#> [1] 4.97
#> 
#> $date1
#> [1] 16578
#> 
#> $date2
#> [1] 18394
#> 

# Easier to view and filter when you split by some variable with many groups
# (note warning when some groups have dbh values full of NA)
recruitment_df(census1, census2, split1 = census1$sp)

#> Warning: At least one split-group contains all `dbh` values equal to NA.
#>   * Consider removing those groups before running this function.

#> # A tibble: 1,152 x 3
#>     split metric    value
#>     <chr>  <chr>    <dbl>
#>  1 acaldi     N2 5.00e+00
#>  2 acaldi      R 2.00e+00
#>  3 acaldi   rate 1.03e-01
#>  4 acaldi  lower 2.54e-02
#>  5 acaldi  upper 3.03e-01
#>  6 acaldi   time 4.95e+00
#>  7 acaldi  date1 1.66e+04
#>  8 acaldi  date2 1.85e+04
#>  9 aegipa     N2 0.00e+00
#> 10 aegipa      R 0.00e+00
#> # ... with 1,142 more rows

# Same for `mortality()`; lets explore a few species
(all_sp <- mortality_df(census1, census2, split1 = census1$sp))

#> Warning: At least one split-group contains all `dbh` values equal to NA.
#>   * Consider removing those groups before running this function.

#> # A tibble: 1,296 x 3
#>     split  metric    value
#>     <chr>   <chr>    <dbl>
#>  1 acaldi       N 3.00e+00
#>  2 acaldi       D 0.00e+00
#>  3 acaldi    rate 0.00e+00
#>  4 acaldi   lower 0.00e+00
#>  5 acaldi   upper 1.87e-01
#>  6 acaldi    time 4.94e+00
#>  7 acaldi   date1 1.66e+04
#>  8 acaldi   date2 1.84e+04
#>  9 acaldi dbhmean 1.30e+01
#> 10 aegipa       N 0.00e+00
#> # ... with 1,286 more rows

few_sp <- sample(unique(all_sp$split), 3)
long_format <- subset(all_sp, split %in% few_sp)
head(long_format, 10)

#> # A tibble: 10 x 3
#>     split  metric    value
#>     <chr>   <chr>    <dbl>
#>  1 coccma       N 2.00e+00
#>  2 coccma       D 0.00e+00
#>  3 coccma    rate 0.00e+00
#>  4 coccma   lower 0.00e+00
#>  5 coccma   upper 2.46e-01
#>  6 coccma    time 4.99e+00
#>  7 coccma   date1 1.65e+04
#>  8 coccma   date2 1.83e+04
#>  9 coccma dbhmean 3.65e+01
#> 10 guarsp       N 2.00e+00

# From long to wide format
library(tidyr)
wide_format <- tidyr::spread(long_format, key = split, value = value)
wide_format

#> # A tibble: 9 x 4
#>    metric   coccma   guarsp   tropra
#> *   <chr>    <dbl>    <dbl>    <dbl>
#> 1       D 0.00e+00 1.00e+00 0.00e+00
#> 2   date1 1.65e+04 1.66e+04 1.66e+04
#> 3   date2 1.83e+04 1.84e+04 1.84e+04
#> 4 dbhmean 3.65e+01 1.40e+02 1.10e+01
#> 5   lower 0.00e+00 1.99e-02 0.00e+00
#> 6       N 2.00e+00 2.00e+00 1.00e+00
#> 7    rate 0.00e+00 1.39e-01 0.00e+00
#> 8    time 4.99e+00 4.98e+00 4.94e+00
#> 9   upper 2.46e-01 4.75e-01 3.73e-01