Convert to quadrat indices.

These functions output quadrat indices from either row and column numbers (to_id_rowcol), or gx and gy coordinates (to_id_gxgy]). They replace ctfs::rowcol.to.index and ctfs::gxgy.to.index, but these old names are kept as aliases to help users find the newer names.

to_id_rowcol(.row, .col, gridsize, plotdim)

to_id_gxgy(gx, gy, gridsize, plotdim)

Arguments

.row, .col
Row and column number.

gridsize
Side of the square quadrate.

plotdim
The x and y dimensions of the plot.

gx, gy
Tree x and y coordinate.

Value

A numeric vector of indices.

Examples

library(dplyr)
mini_cns <- bci::bci12full7 %>%
  select(gx, gy) %>%
  sample_n(10)
to_id_gxgy(mini_cns$gx, mini_cns$gy, gridsize = 20, plotdim = c(1000, 500))

#>  [1]  772  385  832  502  252  268   36  728 1004  626

to_id_rowcol(.row = 1:10, .col = 6:15, gridsize = 20, plotdim = c(1000, 500))

#>  [1] 126 152 178 204 230 256 282 308 334 360