Abundance and basal area.

abundance() and basal_area():

Warn if they detect multiple censusid and multiple plots.

Can compute by groups of data created with group_by().

abundance():

Is similar to dplyr::n().

Warns if it detects duplicated values of treeid.

basal_area():

Warns if it detects duplicated values of stemid.

Does not convert units. For that see convert_unit().

abundance(x)

basal_area(x)

Arguments

x
A dataframe. basal_area() requires a column named dbh (case insensitive).

Warning

To pick specific rows (e.g. to pick "alive" stems or dbh within some range) you should use functions such as [, subset(), or dplyr::filter() before using abundance(). (Notice the difference with ctfs::abundance, which includes arguments to deal with dbh and status).

See also

Examples

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

library(fgeo.tool)

#> 
#> Attaching package: 'fgeo.tool'

#> The following object is masked from 'package:stats':
#> 
#>     filter

# abundance() -------------------------------------------------------------

# Similar to dplyr::n()
abundance(data.frame(1))

#> # A tibble: 1 x 1
#>       n
#>   <int>
#> 1     1

vft <- tibble::tribble(
  ~PlotName, ~CensusID, ~TreeID, ~StemID, ~DBH,
  "p",         1,     "1",   "1.1",   10,
  "q",         2,     "1",   "1.1",   10
)

# * Warns if it detects multiple values of censusid or plotname
# * Also warns if it detects duplicated values of treeid
abundance(vft)

#> Warning: `treeid`: Duplicated values were detected. Do you need to pick main stems?

#> Warning: `censusid`: Multiple values were detected. Do you need to group by censusid?

#> Warning: `plotname`: Multiple values were detected. Do you need to pick a single plot?

#> # A tibble: 1 x 1
#>       n
#>   <int>
#> 1     2

# You should probably work with a single plotname. 
# Yet your data may have multiple stems per treeid and even multiple measures
# per stemid (when trees have buttressess).
vft2 <- tibble::tribble(
  ~CensusID, ~TreeID, ~StemID, ~DBH, ~HOM,
  1,     "1",   "1.1",   88,  130,
  1,     "1",   "1.1",   10,  160,
  1,     "2",   "2.1",   20,  130,
  1,     "2",   "2.2",   30,  130,
)

# You should count only the main stem of each tree
(main_stem <- pick_main_stem(vft2))

#> # A tibble: 2 x 5
#>   CensusID TreeID StemID   DBH   HOM
#>      <dbl> <chr>  <chr>  <dbl> <dbl>
#> 1        1 1      1.1       10   160
#> 2        1 2      2.2       30   130

abundance(main_stem)

#> # A tibble: 1 x 1
#>       n
#>   <int>
#> 1     2

vft3 <- tibble::tribble(
  ~CensusID, ~TreeID, ~StemID, ~DBH, ~HOM,
  1,     "1",   "1.1",   20,  130,
  1,     "1",   "1.2",   10,  160,  # Main stem
  2,     "1",   "1.1",   12,  130,
  2,     "1",   "1.2",   22,  130   # Main stem
)

# You can compute by groups
by_census <- group_by(vft3, CensusID)
(main_stems_by_census <- pick_main_stem(by_census))

#> # A tibble: 2 x 5
#> # Groups:   CensusID [2]
#>   CensusID TreeID StemID   DBH   HOM
#>      <dbl> <chr>  <chr>  <dbl> <dbl>
#> 1        1 1      1.2       10   160
#> 2        2 1      1.2       22   130

abundance(main_stems_by_census)

#> # A tibble: 2 x 2
#> # Groups:   CensusID [2]
#>   CensusID     n
#>      <dbl> <int>
#> 1        1     1
#> 2        2     1

# basal_area() ------------------------------------------------------------

# Data must have a column named dbh (case insensitive)
basal_area(data.frame(dbh = 1))

#> # A tibble: 1 x 1
#>   basal_area
#>        <dbl>
#> 1      0.785

# * Warns if it detects multiple values of censusid or plotname
# * Also warns if it detects duplicated values of stemid
basal_area(vft)

#> Warning: `stemid`: Duplicated values were detected. Do you need to pick largest `hom` values?

#> Warning: `censusid`: Multiple values were detected. Do you need to group by censusid?

#> Warning: `plotname`: Multiple values were detected. Do you need to pick a single plot?

#> # A tibble: 1 x 1
#>   basal_area
#>        <dbl>
#> 1       157.

# First you may pick the main stemid of each stem
(main_stemids <- pick_main_stemid(vft2))

#> # A tibble: 3 x 5
#>   CensusID TreeID StemID   DBH   HOM
#>      <dbl> <chr>  <chr>  <dbl> <dbl>
#> 1        1 1      1.1       10   160
#> 2        1 2      2.1       20   130
#> 3        1 2      2.2       30   130

basal_area(main_stemids)

#> # A tibble: 1 x 1
#>   basal_area
#>        <dbl>
#> 1      1100.

# You can compute by groups
basal_area(by_census)

#> # A tibble: 2 x 2
#> # Groups:   CensusID [2]
#>   CensusID basal_area
#>      <dbl>      <dbl>
#> 1        1       393.
#> 2        2       493.

# To convert units see ?convert_unit()
ba <- basal_area(by_census)
convert_unit_at(ba, .at = "basal_area", from = "mm2", to = "hectare")

#> # A tibble: 2 x 2
#> # Groups:   CensusID [2]
#>   CensusID   basal_area
#>      <dbl>        <dbl>
#> 1        1 0.0000000393
#> 2        2 0.0000000493