Count individuals by group.

To calculate the abundance of individuals by groups is a particular case of the more general issue of counting observations. This is such a common action that there are great tools for doing just that. Some of the most widely used and well tested are count(), tally(), add_count() and add_tally() from the dplyr package. For convenience, those functions are reexported by by fgeo.abundance so they will be available to you when you run library(fgeo.abundance). The examples below are show specifically how these functions apply to ForestGEO data. To learn more about these functions, and to see more general examples, please refer to the original documentation of dplyr::count().

See also

Examples

census <- tibble::tibble(
  stemID = 1:6,
  quadrat = paste0("000", rep(1:2, each = 3)),
  sp = c(paste0("sp", c(1, 1, 2)), paste0("sp", c(3, 3, 3))),
  dbh = abs(sample(rnorm(100), 6) * 10)
)
census

#> # A tibble: 6 x 4
#>   stemID quadrat sp        dbh
#>    <int> <chr>   <chr>   <dbl>
#> 1      1 0001    sp1    1.12  
#> 2      2 0001    sp1    0.381 
#> 3      3 0001    sp2    3.63  
#> 4      4 0002    sp3    0.0557
#> 5      5 0002    sp3   16.3   
#> 6      6 0002    sp3   13.0   

# Abundance (rows count) by quadrat
count(census, quadrat)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        3
#> 2 0002        3

# Abundance (rows count) by species
count(census, sp)

#> # A tibble: 3 x 2
#>   sp        n
#>   <chr> <int>
#> 1 sp1       2
#> 2 sp2       1
#> 3 sp3       3

# Abundance (rows count) by quadrat by species
count(census, quadrat, sp)

#> # A tibble: 3 x 3
#>   quadrat sp        n
#>   <chr>   <chr> <int>
#> 1 0001    sp1       2
#> 2 0001    sp2       1
#> 3 0002    sp3       3

# Richness by quadrat:
# Count is designed so that you can call it repeatedly, each time rolling up
# a level of detail.
# Now, each row by quadrat is a unique species, so counting rows gives
# richness
n_quad_sp <- count(census, quadrat, sp)
richness_quad <- count(n_quad_sp, quadrat)
richness_quad

#> # A tibble: 2 x 2
#>   quadrat    nn
#>   <chr>   <int>
#> 1 0001        2
#> 2 0002        1

# Singleton:
# add_count() is useful for groupwise filtering e.g.: show only species that
# have a single member
out <- add_count(census, quadrat, sp)
subset(out, n == 1)

#> # A tibble: 1 x 5
#>   stemID quadrat sp      dbh     n
#>    <int> <chr>   <chr> <dbl> <int>
#> 1      3 0001    sp2    3.63     1