Count number of individuals in total or by groups.

Count number of individuals in total (abundance_tally()) or by groups (abundance()).

abundance(x, group_by = c("quadrat", "sp"), only_alive = TRUE)

abundance_tally(x, only_alive = TRUE)

Arguments

x
Census data.

group_by
Character string giving the names of variables to group by.

only_alive
If TRUE, counts only alive individuals (`status == "A"`). If FALSE counts individuals of all available statuses.

Value

abundance_tally(): An integer giving the number of rows in x, which for census data equals the number of unique individuals.

abundance(): a data frame giving the number of individuals in each group defined by group_by.

See also

dplyr::tally(), dplyr::count().

Other grouped summaries: basal_area

Examples

stem <- bci12s7mini

# Returns a data frame by groups
n <- abundance(stem)
head(n)

#>   quadrat     sp n
#> 1    0000 soroaf 1
#> 2    0002 hybapr 1
#> 3    0006 hybapr 2
#> 4    0009 faraoc 1
#> 5    0009 hybapr 1
#> 6    0020 tet2pa 1

# Instead of a data frame, return an integer representing a tally
abundance_tally(stem)

#> [1] 665

# Count not only alive
n <- abundance(stem, group_by = c("status", "sp"), only_alive = FALSE)
ordered_by_species <- order(n$sp)
head(n[ordered_by_species, ])

#>     status     sp  n
#> 1        A acaldi  5
#> 112      D acaldi 11
#> 222      G acaldi 11
#> 113      D aegipa  1
#> 114      D alchco  1
#> 223      G alchco 10

# Alternatives ----------------------------------------------------------

library(dplyr)

alive <- filter(stem, status == "A")

grouped <- group_by(alive, quadrat, sp)
summarise(grouped, n = n())

#> # A tibble: 568 x 3
#> # Groups:   quadrat [?]
#>    quadrat     sp     n
#>      <chr>  <chr> <int>
#>  1    0000 soroaf     1
#>  2    0002 hybapr     1
#>  3    0006 hybapr     2
#>  4    0009 faraoc     1
#>  5    0009 hybapr     1
#>  6    0020 tet2pa     1
#>  7    0022 sympgl     1
#>  8    0100 des2pa     1
#>  9    0100 psycho     1
#> 10    0101 ingafa     1
#> # ... with 558 more rows

# Or
count(alive, quadrat, sp)

#> # A tibble: 568 x 3
#>    quadrat     sp     n
#>      <chr>  <chr> <int>
#>  1    0000 soroaf     1
#>  2    0002 hybapr     1
#>  3    0006 hybapr     2
#>  4    0009 faraoc     1
#>  5    0009 hybapr     1
#>  6    0020 tet2pa     1
#>  7    0022 sympgl     1
#>  8    0100 des2pa     1
#>  9    0100 psycho     1
#> 10    0101 ingafa     1
#> # ... with 558 more rows