Basal area of each individual.

basal_area() is a generic function with methods for numeric vectors and dataframes. To calculate basal area within groups, feed basal_area() with a dataframe previously grouped with group_by() (see examples and section See also). group_by() comes from dplyr but for convenience it is reexported by fgeo.abundance.

basal_area(x, ...)

# S3 method for default
basal_area(x, ...)

# S3 method for data.frame
basal_area(x, dbh = dbh, ...)

# S3 method for numeric
basal_area(x, ...)

Arguments

x
Numeric vector or dataframe giving the individuals' diameter.

...
Other arguments passed to methods.

dbh
If x is a dataframe, dbh is the bare name of the column giving individual's diameter.

Value

Each individual calculation is simply the area of a circle given its diameter; thus the unit of the returned value is the square of the input units unit (i.e. for input in mm, output is mm^2; for input in cm, output is cm^2, and so on).

basal_area.numeric(): A numeric vector giving the basal area of each individual.

See also

Examples

library(dplyr)

stem <- fgeo.data::luquillo_stem_random_tiny

# Calculate basal area on a vector
ba <- basal_area(stem$dbh)
ba

#>  [1] 18869.19088 22167.07776 10935.88403  1995.03700  2042.82062   254.46900
#>  [7]   314.15927   397.60782   593.95736   113.09734   572.55526 18869.19088
#> [13] 23506.18163 16512.99639  1809.55737  3959.19214   128.67964   651.44065
#> [19]   158.36769 20867.24380 23506.18163 16971.66891  2375.82944  9503.31778
#> [25]   134.78218   897.27028   463.76976   387.07563   136.84778   235.06182
#> [31]   206.11989 16286.01632 11499.01451 18869.19088 26015.52876  2375.82944
#> [37] 15174.67792   897.27028   419.09631   165.12996   506.70748   237.78715
#> [43] 16741.54725 14526.72443    84.94867  5944.67870 17907.86352 19359.27933
#> [49] 23235.21927 17203.36137  2375.82944 19113.44970   855.29860   510.70516
#> [55]   237.78715 16741.54725 14957.12262  6082.12338 19359.27933 23506.18163
#> [61] 17203.36137 20106.19298 16741.54725 15614.50089  5944.67870          NA
#> [67]          NA          NA          NA          NA          NA          NA
#> [73]          NA          NA          NA          NA          NA

# Add result to the original dataframe
stem$ba <- ba
# Same
stem <- mutate(stem, ba = basal_area(dbh))
# Reordering columns to show `ba` first
select(stem, ba, everything())

#> # A tibble: 77 x 20
#>        ba treeID stemID tag   StemTag sp     quadrat    gx    gy MeasureID
#>     <dbl>  <int>  <int> <chr> <chr>   <chr>  <chr>   <dbl> <dbl>     <int>
#>  1 18869.  10775  13496 1128  1128    PREMON 402      79.3  39.8     13496
#>  2 22167.  10863  13603 1129  1129    PREMON 402      79.0  36.5     13603
#>  3 10936.  23710  29640 14434 14434   PREMON 517      82.7 330.      29640
#>  4  1995.  25331  31937 16164 16164   SLOBER 402      70.8  21.0     31937
#>  5  2043.  63997  77760 75062 75062   CASARB 517      86.0 340.      77760
#>  6   254.  63997  77761 75062 75063   CASARB 517      86.0 340.      77761
#>  7   314.  63997  77762 75062 75064   CASARB 517      86.0 340.      77762
#>  8   398.  63997  77763 75062 75065   CASARB 517      86.0 340.      77763
#>  9   594.  63997  77764 75062 75066   CASARB 517      86.0 340.      77764
#> 10   113.  64070  77850 75147 75147   CASARB 517      99.8 325.      77850
#> # ... with 67 more rows, and 10 more variables: CensusID <int>, dbh <dbl>,
#> #   pom <chr>, hom <dbl>, ExactDate <dbl>, DFstatus <chr>, codes <chr>,
#> #   countPOM <dbl>, status <chr>, date <dbl>

# Calculate basal area on the entire dataframe
basal_area(stem)

#> # A tibble: 1 x 1
#>   basal_area
#>        <dbl>
#> 1    586336.

# Now by groups
basal_area(group_by(stem, quadrat))

#> # A tibble: 2 x 2
#>   quadrat basal_area
#>   <chr>        <dbl>
#> 1 402        340213.
#> 2 517        246123.

grouped <- group_by(stem, quadrat, sp)
basal_area(grouped)

#> # A tibble: 4 x 3
#> # Groups:   quadrat [?]
#>   quadrat sp     basal_area
#>   <chr>   <chr>       <dbl>
#> 1 402     PREMON    324640.
#> 2 402     SLOBER     15573.
#> 3 517     CASARB     74819.
#> 4 517     PREMON    171304.

# Also works on ViewFullTable
vft <- fgeo.data::luquillo_vft_4quad
head(basal_area(vft$DBH))

#> [1]        NA  26880.25        NA  91326.88 107521.01 110446.62

grouped_vft <- group_by(vft, CensusID, QuadratName)
basal_area(grouped_vft, dbh = DBH)

#> # A tibble: 12 x 3
#> # Groups:   CensusID [?]
#>    CensusID QuadratName basal_area
#>       <int> <chr>            <dbl>
#>  1        4 621           1449189.
#>  2        4 622           1431360.
#>  3        4 721           1401168.
#>  4        4 722           1469214.
#>  5        5 621           1306077.
#>  6        5 622           1650672.
#>  7        5 721           1359760.
#>  8        5 722           1377121.
#>  9        6 621           1205149.
#> 10        6 622           1345448.
#> 11        6 721           1191424.
#> 12        6 722           1268563.