Basal area.
Source: R/basal_area.R

Basal area of each individual (basal_area_ind()).

basal_area(x, group_by = c("quadrat", "sp"), only_alive = TRUE)

basal_area_ind(diameter)

Arguments

x
Census data.

group_by
Character string giving the names of variables to group by.

only_alive
If TRUE, counts only alive individuals (`status == "A"`). If FALSE counts individuals of all available statuses.

diameter
A vector giving the individuals' diameter (generally, x$dbh).

Value

basal_area_ind(): A numeric vector giving the basal area of each individual in the same units as the input vector. This is simply the area of a circle given its diameter.

Details

Units of returned values are the square of the input units. For example, if you input diameter measured in mm, the output will be an area measured in mm^2.

See also

Other grouped summaries: abundance

Examples

stem <- bciex::bci12s7mini

head(stem$dbh)

#> [1] 720  NA  NA  NA  NA  NA

head(basal_area_ind(stem$dbh))

#> Units of returned values are the square of the input units.

#> [1] 407150     NA     NA     NA     NA     NA

head(basal_area(stem))

#> Units of returned values are the square of the input units.

#>   quadrat     sp basal_area
#> 1    0000 soroaf       2642
#> 2    0002 hybapr         95
#> 3    0006 hybapr        475
#> 4    0009 faraoc       1810
#> 5    0009 hybapr          0
#> 6    0020 tet2pa       3117

# Silent
head(suppressMessages(basal_area(stem)))

#>   quadrat     sp basal_area
#> 1    0000 soroaf       2642
#> 2    0002 hybapr         95
#> 3    0006 hybapr        475
#> 4    0009 faraoc       1810
#> 5    0009 hybapr          0
#> 6    0020 tet2pa       3117