Basal area of each individual.

basal_area() is a generic function with methods for numeric vectors and dataframes. To calculate basal area within groups, feed basal_area() with a dataframe previously grouped with group_by() (see examples and section See also). group_by() comes from dplyr but for convenience it is reexported by fgeo.abundance. add_basal_area() adds a column basal_area

basal_area(x, ...)

# S3 method for default
basal_area(x, ...)

# S3 method for data.frame
basal_area(x, dbh = dbh, ...)

# S3 method for numeric
basal_area(x, ...)

add_basal_area(x, ..., dbh = dbh)

Arguments

x
Numeric vector or dataframe giving the individuals' diameter.

...
Other arguments passed to methods.

dbh
If x is a dataframe, dbh is the bare name of the column giving individual's diameter.

Value

Each individual calculation is simply the area of a circle given its diameter; thus the unit of the returned value is the square of the input units unit (i.e. for input in mm, output is mm^2; for input in cm, output is cm^2, and so on).

basal_area.numeric(): A numeric vector giving the basal area of each individual.

See also

Examples

library(dplyr)

stem <- tibble::tibble(
  stemID = 1:6,
  quadrat = paste0("000", rep(1:2, each = 3)),
  sp = c(paste0("sp", c(1, 1, 2)), paste0("sp", c(3, 3, 3))),
  dbh = abs(sample(rnorm(100), 6) * 10)
)
stem

#> # A tibble: 6 x 4
#>   stemID quadrat sp        dbh
#>    <int> <chr>   <chr>   <dbl>
#> 1      1 0001    sp1    1.12  
#> 2      2 0001    sp1    0.381 
#> 3      3 0001    sp2    3.63  
#> 4      4 0002    sp3    0.0557
#> 5      5 0002    sp3   16.3   
#> 6      6 0002    sp3   13.0   

basal_area(stem$dbh)

#> [1] 9.858736e-01 1.140265e-01 1.034633e+01 2.437816e-03 2.089258e+02
#> [6] 1.324043e+02

add_basal_area(stem)

#> # A tibble: 6 x 5
#>   stemID quadrat sp        dbh basal_area
#>    <int> <chr>   <chr>   <dbl>      <dbl>
#> 1      1 0001    sp1    1.12         353.
#> 2      2 0001    sp1    0.381        353.
#> 3      3 0001    sp2    3.63         353.
#> 4      4 0002    sp3    0.0557       353.
#> 5      5 0002    sp3   16.3          353.
#> 6      6 0002    sp3   13.0          353.

# Sum across all rows
basal_area(stem)

#> # A tibble: 1 x 1
#>   basal_area
#>        <dbl>
#> 1       353.

# Sum across rows by groups
basal_area(
  group_by(stem, quadrat, sp)
)

#> # A tibble: 3 x 3
#> # Groups:   quadrat [?]
#>   quadrat sp    basal_area
#>   <chr>   <chr>      <dbl>
#> 1 0001    sp1         1.10
#> 2 0001    sp2        10.3 
#> 3 0002    sp3       341.  

# Same but returns all rows
add_basal_area(stem, quadrat, sp)

#> # A tibble: 6 x 5
#>   stemID quadrat sp        dbh basal_area
#>    <int> <chr>   <chr>   <dbl>      <dbl>
#> 1      1 0001    sp1    1.12         1.10
#> 2      2 0001    sp1    0.381        1.10
#> 3      3 0001    sp2    3.63        10.3 
#> 4      4 0002    sp3    0.0557     341.  
#> 5      5 0002    sp3   16.3        341.  
#> 6      6 0002    sp3   13.0        341.  

# Dealing with missing values
stem2 <- stem
stem2[6, 4] <- NA
stem2

#> # A tibble: 6 x 4
#>   stemID quadrat sp        dbh
#>    <int> <chr>   <chr>   <dbl>
#> 1      1 0001    sp1    1.12  
#> 2      2 0001    sp1    0.381 
#> 3      3 0001    sp2    3.63  
#> 4      4 0002    sp3    0.0557
#> 5      5 0002    sp3   16.3   
#> 6      6 0002    sp3   NA     

# basal_area() of missing dbh in a vector results in NA
basal_area(stem2$dbh)

#> [1] 9.858736e-01 1.140265e-01 1.034633e+01 2.437816e-03 2.089258e+02
#> [6]           NA

# add_basal_area() of missing dbh in a dataframe results in NA
# Same if using add_basal_area()
add_basal_area(stem2)

#> # A tibble: 6 x 5
#>   stemID quadrat sp        dbh basal_area
#>    <int> <chr>   <chr>   <dbl>      <dbl>
#> 1      1 0001    sp1    1.12         220.
#> 2      2 0001    sp1    0.381        220.
#> 3      3 0001    sp2    3.63         220.
#> 4      4 0002    sp3    0.0557       220.
#> 5      5 0002    sp3   16.3          220.
#> 6      6 0002    sp3   NA            220.

# But basal_area() of missing dbh in a dataframe ignores NA's
basal_area(stem2)

#> # A tibble: 1 x 1
#>   basal_area
#>        <dbl>
#> 1       220.

basal_area(group_by(stem2, quadrat, sp))

#> # A tibble: 3 x 3
#> # Groups:   quadrat [?]
#>   quadrat sp    basal_area
#>   <chr>   <chr>      <dbl>
#> 1 0001    sp1         1.10
#> 2 0001    sp2        10.3 
#> 3 0002    sp3       209.  

# Missing dbh is ignored, and `basal_area` refers to the group not each row.
add_basal_area(stem2, quadrat, sp)

#> # A tibble: 6 x 5
#>   stemID quadrat sp        dbh basal_area
#>    <int> <chr>   <chr>   <dbl>      <dbl>
#> 1      1 0001    sp1    1.12         1.10
#> 2      2 0001    sp1    0.381        1.10
#> 3      3 0001    sp2    3.63        10.3 
#> 4      4 0002    sp3    0.0557     209.  
#> 5      5 0002    sp3   16.3        209.  
#> 6      6 0002    sp3   NA          209.