Create tables of abundance and basal area by (round mean) year.

Create tables of abundance and basal area by (round mean) year.

byyr_abundance(vft)

byyr_basal_area(vft)

Arguments

vft
A dataframe; particularly a ForestGEO ViewFullTable.

Value

A dataframe.

Examples

library(fgeo.base)
library(fgeo.tool)

vft <- data.frame(
  PlotName = "p",
  PlotCensusNumber = c(1, 1, 2, 2, 1, 2),
  CensusID = c(1, 1, 2, 2, 1, 2),
  ExactDate = c(
    "2001-01-01", "2001-01-01", "2002-01-01", "2002-01-01",
    "2001-01-01", "2002-01-01"
  ),
  Tag = c("0001", "0001", "0001", "0001", "0002", "0002"),
  TreeID = c("0001", "0001", "0001", "0001", "0002", "0002"),
  SteemID = c(  "1",    "2",    "1",    "2",    "1",    "1"),
  Status = c("alive", "alive", "alive", "dead", "alive", "dead"),
  DBH = c(11, 15, 12, NA, 21, NA),
  Genus = c("A", "A", "A", "A", "B", "B"),
  SpeciesName = c("a", "a", "a", "a", "b", "b"),
  Family = "f",
  stringsAsFactors = FALSE
)
vft

#>   PlotName PlotCensusNumber CensusID  ExactDate  Tag TreeID SteemID Status DBH
#> 1        p                1        1 2001-01-01 0001   0001       1  alive  11
#> 2        p                1        1 2001-01-01 0001   0001       2  alive  15
#> 3        p                2        2 2002-01-01 0001   0001       1  alive  12
#> 4        p                2        2 2002-01-01 0001   0001       2   dead  NA
#> 5        p                1        1 2001-01-01 0002   0002       1  alive  21
#> 6        p                2        2 2002-01-01 0002   0002       1   dead  NA
#>   Genus SpeciesName Family
#> 1     A           a      f
#> 2     A           a      f
#> 3     A           a      f
#> 4     A           a      f
#> 5     B           b      f
#> 6     B           b      f

# First pick the data you want
pick1 <- fgeo.base::pick_plotname(vft, "p")

#> Using: p.

pick2 <- drop_dead_trees_by_cns(pick1)

#> Calculating tree-status (from stem `Status`) by `PlotCensusNumber`.

#> Dropping trees which status is dead.

pick3 <- pick_dbh_min(pick2, 10)

byyr_abundance(pick3)

#> # A tibble: 2 x 4
#>   species Family `2001` `2002`
#>   <chr>   <chr>   <dbl>  <dbl>
#> 1 A a     f           1      1
#> 2 B b     f           1      0

ba <- byyr_basal_area(pick3)
ba

#> # A tibble: 2 x 4
#>   species Family `2001` `2002`
#>   <chr>   <chr>   <dbl>  <dbl>
#> 1 A a     f        272.   113.
#> 2 B b     f        346.     0 

# Convert units and standardize by plot size in hectares
years <- c("2000", "2001")
in_he <- convert_unit_at(ba, .at = years, from = "mm2", to = "hectare")
standardize_at(in_he, .at = years, denominator = 50)

#> # A tibble: 2 x 4
#>   species Family   `2001` `2002`
#>   <chr>   <chr>     <dbl>  <dbl>
#> 1 A a     f      5.43e-10   113.
#> 2 B b     f      6.93e-10     0