Count distinct occurrences of a variable by groups (tree or stem abundance).

These functions are usually used on grouped data created by dplyr::group_by() (reexported by fgeo). The output will have one row for each group.

count_distinct() counts distinct occurrences of a variable.

abundance_stem() and abundance_tree() are simpler wrappers for the specifically counting distinct occurrences of the variables stemID and treeID which uniquely identify each stem and tree in census datasets of ForestGEO (both stem and tree tables).

count_distinct(.data, .var)

abundance_stem(.data)

abundance_tree(.data)

Arguments

.data
A dataframe, commonly grouped with group_by().

.var
A variable to count distinct occurrences.

Value

An object of the same class as .data. One grouping level will be dropped.

See also

Examples

library(dplyr)

# Example data
census <- tibble::tibble(
  treeID = c(1, 1, 2, 3, 3, 3),
  stemID = c(1, 2, 3, 4, 5, 6),
  quadrat = paste0("000", rep(1:2, each = 3)),
  sp = c(paste0("sp", c(1, 1, 2)), paste0("sp", c(3, 3, 3))),
  dbh = abs(sample(rnorm(100), 6) * 10)
)
census

#> # A tibble: 6 x 5
#>   treeID stemID quadrat sp      dbh
#>    <dbl>  <dbl> <chr>   <chr> <dbl>
#> 1      1      1 0001    sp1   18.8 
#> 2      1      2 0001    sp1    1.08
#> 3      2      3 0001    sp2    7.96
#> 4      3      4 0002    sp3    8.95
#> 5      3      5 0002    sp3    5.58
#> 6      3      6 0002    sp3   15.3 

count_distinct(census, stemID)

#> # A tibble: 1 x 1
#>       n
#>   <int>
#> 1     6

by_quad <- group_by(census, quadrat)
count_distinct(by_quad, stemID)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        3
#> 2 0002        3

count_distinct(by_quad, treeID)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        2
#> 2 0002        1

abundance_stem(by_quad)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        3
#> 2 0002        3

abundance_tree(by_quad)

#> # A tibble: 2 x 2
#>   quadrat     n
#>   <chr>   <int>
#> 1 0001        2
#> 2 0002        1

by_sp <- group_by(census, sp)
abundance_stem(by_sp)

#> # A tibble: 3 x 2
#>   sp        n
#>   <chr> <int>
#> 1 sp1       2
#> 2 sp2       1
#> 3 sp3       3

abundance_tree(by_sp)

#> # A tibble: 3 x 2
#>   sp        n
#>   <chr> <int>
#> 1 sp1       1
#> 2 sp2       1
#> 3 sp3       1

by_quad_sp <- group_by(census, quadrat, sp)
abundance_stem(by_quad_sp)

#> # A tibble: 3 x 3
#> # Groups:   quadrat [?]
#>   quadrat sp        n
#>   <chr>   <chr> <int>
#> 1 0001    sp1       2
#> 2 0001    sp2       1
#> 3 0002    sp3       3

abundance_tree(by_quad_sp)

#> # A tibble: 3 x 3
#> # Groups:   quadrat [?]
#>   quadrat sp        n
#>   <chr>   <chr> <int>
#> 1 0001    sp1       1
#> 2 0001    sp2       1
#> 3 0002    sp3       1