Annual growth rates accross entire census or within groups.

Annual growth rates accross entire census or within groups.

growth(census1, census2, rounddown = FALSE, method = "I", stdev = FALSE,
  dbhunit = "mm", mindbh = 10, growthcol = "dbh", err.limit = 4,
  maxgrow = 75, split1 = NULL, split2 = NULL)

growth_df(census1, census2, rounddown = FALSE, method = "I",
  stdev = FALSE, dbhunit = "mm", mindbh = 10, growthcol = "dbh",
  err.limit = 4, maxgrow = 75, split1 = NULL)

Arguments

census1
The R Analytical Table for a single census, either tree or stem.

census2
The matching R Analytical Table for a later census.

rounddown
If TRUE, all dbh < 55 are rounded down to the nearest multiple of 5.

method
Use 'I' to calculate annual dbh increment: (dbh2 - dbh1)/time, or 'E' to calculate the relative growth rate (log(dbh2) - log(dbh1))/time.

stdev
Logical. Default (FALSE) returns confidence limits, otherwise returns the SD in growth rate per group.

dbhunit
'cm' or 'mm'.

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

growthcol
defines how growth is measured, either 'dbh'or 'agb'(agb=biomass)

err.limit
A number. Numbers such as 10000 are high and will return all measures.

maxgrow
A number. Numbers such as 10000 are high and will return all measures.

split1
a vector of categories, one per individual

split2
another vector of categories, one per individual

Value

growth() retruns a list with components:

rate, the mean annualized growth rate per category selected, either dbh increment, or relative growth

N, the number of individuals included in the mean (not counting any excluded)

clim (or sd with stdev = TRUE), width of confidence interval; add this number to the mean rate to get upper confidence limit, substract to get lower

dbhmean, mean dbh in census 1 of individuals included

time, mean time interval in years

date1, mean date included individuals were measured in census 1, as julian object (R displays as date, but treats as integer)

date2, mean date in census 2.

Details

growth() also requires the columns status and stemID in order to determine which stems should have dbh change calculated.

See also

Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: CI, IfElse, addBinParam, arrangeParam.Gibbs, arrangeParam.llike, bad.binparam, binGraphManySpecies.Panel, binGraphManySpecies, binGraphSampleSpecies, biomass.growth, calcMortIndivTable, calculateBinModel.AIC, calculateBinModel.BIC, colMedians, compare.growthbinmodel, defineBinBreaks, defineSDpar, dgammadexp, drp, enoughSamplePerBin, extract.growthdata, fill.dimension, find.climits, graph.growthmodel.spp, graph.growthmodel, graph.outliers.spp, graph.outliers, growth.biomass.indiv, growth.dbh, growth.eachspp, growth.flexbin, growth.indiv, individual_grow.table, linear.model.ctr, linear.model, linearmodel.bin.set, linearmodel.bin, llike.GaussModelSD, llike.GaussModel, llike.linearbin.optim, logit, metrop1step, model.xy, mortality.calculation, mortality.dbh, mortality.eachspp, mortality, overlay.growthbinmodel, pst, pts.to.interceptslope, rndown5, run.growthbin.manyspp, run.growthfit.bin, trim.growth, wideEnoughBins

Other functions to compare two censuses: mortality

Examples

# Filtering only 200 trees to make the data small
few_trees <- sample(unique(bci12s6mini$tag), 200)
census1 <- subset(bci12s6mini, tag %in% few_trees)
census2 <- subset(bci12s7mini, tag %in% few_trees)

# Across entire dataset
growth(census1, census2)

#> $rate
#> [1] 1.23
#> 
#> $N
#> [1] 78
#> 
#> $clim
#> [1] 0.66
#> 
#> $dbhmean
#> [1] 55.1
#> 
#> $time
#> [1] 4.96
#> 
#> $date1
#> [1] 16584
#> 
#> $date2
#> [1] 18396
#> 

# Within groups defined by one "splitting" variable
# Note warning if split variable has grups which dbh is full of NA
growth_metrics <- growth(census1, census2, split1 = census1$sp)

#> Warning: At least one split-group contains all `dbh` values equal to NA.
#>   * Consider removing those groups before running this function.

# View just a few species with `head()`
lapply(growth_metrics, head)

#> $rate
#> acaldi alibed alsebl beilpe calolo cassel 
#>  0.402     NA  1.209  1.014     NA  1.033 
#> 
#> $N
#> acaldi alibed alsebl beilpe calolo cassel 
#>      1      0      2      2      0      1 
#> 
#> $clim
#> acaldi alibed alsebl beilpe calolo cassel 
#>     NA     NA  5.201  0.907     NA     NA 
#> 
#> $dbhmean
#> acaldi alibed alsebl beilpe calolo cassel 
#>     11     NA     67    113     NA     33 
#> 
#> $time
#> acaldi alibed alsebl beilpe calolo cassel 
#>   4.97     NA   4.96   4.94     NA   4.84 
#> 
#> $date1
#> acaldi alibed alsebl beilpe calolo cassel 
#>  16625     NA  16612  16634     NA  16554 
#> 
#> $date2
#> acaldi alibed alsebl beilpe calolo cassel 
#>  18442     NA  18423  18438     NA  18322 
#> 

# Within groups defined by two "splitting" variables
# Also warnings here (one warning per problematic splitting variable)
split_by_two <- growth(census1, census2,
  split1 = census1$sp, split2 = census1$quadrat)

#> Warning: At least one split-group contains all `dbh` values equal to NA.
#>   * Consider removing those groups before running this function.

#> Warning: At least one split-group contains all `dbh` values equal to NA.
#>   * Consider removing those groups before running this function.

lapply(split_by_two, function(x) head(x[1:6]))

#> $rate
#> [1] NA NA NA NA NA NA
#> 
#> $N
#> [1] 0 0 0 0 0 0
#> 
#> $clim
#> [1] NA NA NA NA NA NA
#> 
#> $dbhmean
#> [1] NA NA NA NA NA NA
#> 
#> $time
#> [1] NA NA NA NA NA NA
#> 
#> $date1
#> [1] NA NA NA NA NA NA
#> 
#> $date2
#> [1] NA NA NA NA NA NA
#> 

str(split_by_two)

#> List of 7
#>  $ rate   : num [1:69, 1:184] NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:69] "acaldi" "alibed" "alsebl" "beilpe" ...
#>   .. ..$ : chr [1:184] "0019" "0100" "0101" "0118" ...
#>  $ N      : num [1:69, 1:184] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:69] "acaldi" "alibed" "alsebl" "beilpe" ...
#>   .. ..$ : chr [1:184] "0019" "0100" "0101" "0118" ...
#>  $ clim   : num [1:69, 1:184] NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:69] "acaldi" "alibed" "alsebl" "beilpe" ...
#>   .. ..$ : chr [1:184] "0019" "0100" "0101" "0118" ...
#>  $ dbhmean: num [1:69, 1:184] NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:69] "acaldi" "alibed" "alsebl" "beilpe" ...
#>   .. ..$ : chr [1:184] "0019" "0100" "0101" "0118" ...
#>  $ time   : num [1:69, 1:184] NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:69] "acaldi" "alibed" "alsebl" "beilpe" ...
#>   .. ..$ : chr [1:184] "0019" "0100" "0101" "0118" ...
#>  $ date1  : num [1:69, 1:184] NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:69] "acaldi" "alibed" "alsebl" "beilpe" ...
#>   .. ..$ : chr [1:184] "0019" "0100" "0101" "0118" ...
#>  $ date2  : num [1:69, 1:184] NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:69] "acaldi" "alibed" "alsebl" "beilpe" ...
#>   .. ..$ : chr [1:184] "0019" "0100" "0101" "0118" ...

# Use other arguments carefully; read `?growth()`

# Calculate not annual growth rate but relative growth rate
growth(census1, census2, method = "E")

#> $rate
#> [1] 0.0228
#> 
#> $N
#> [1] 78
#> 
#> $clim
#> [1] 0.00723
#> 
#> $dbhmean
#> [1] 55.1
#> 
#> $time
#> [1] 4.96
#> 
#> $date1
#> [1] 16584
#> 
#> $date2
#> [1] 18396
#> 

# Return `sd` instead of `clim`
growth(census1, census2, stdev = TRUE)

#> $rate
#> [1] 1.23
#> 
#> $N
#> [1] 78
#> 
#> $sd
#> [1] 2.93
#> 
#> $dbhmean
#> [1] 55.1
#> 
#> $time
#> [1] 4.96
#> 
#> $date1
#> [1] 16584
#> 
#> $date2
#> [1] 18396
#> 

# Include all living trees
growth(census1, census2, mindbh = NULL)

#> $rate
#> [1] 1.23
#> 
#> $N
#> [1] 78
#> 
#> $clim
#> [1] 0.66
#> 
#> $dbhmean
#> [1] 55.1
#> 
#> $time
#> [1] 4.96
#> 
#> $date1
#> [1] 16584
#> 
#> $date2
#> [1] 18396
#> 

# Measure growth not based on `dbh` but on `agb`
growth(census1, census2, growthcol = "agb")

#> $rate
#> [1] 0.00333
#> 
#> $N
#> [1] 78
#> 
#> $clim
#> [1] 0.00375
#> 
#> $dbhmean
#> [1] 55.1
#> 
#> $time
#> [1] 4.96
#> 
#> $date1
#> [1] 16584
#> 
#> $date2
#> [1] 18396
#>