Count and basal area of neighbouring stems.

Includes all values of status. You shouild pick the values you want before using these functions (e.g. to drop dead stems).

neighbor_count(.data, .subset = NULL, r, plotdim = NULL)

neighbor_basal_area(.data, .subset = NULL, r, plotdim = NULL)

Arguments

.data
A Dataframe; particularly a ForestGEO census table (tree or stem).

.subset
An optional dataframe giving the (gx, gy) coordinates of specific individuals (tag and sp) for which to count neighbors. For example, .subset may be a subset of only one species; or seedling that are not part of census .data. .subset must have columns gx, gy, sp, and tag.

r
Radius.

plotdim
The x and y dimensions of the plot.

Value

A dataframe.

Examples

tree <- fgeo.data::luquillo_tree5_random
neighbor_count(tree, r = 20, plotdim = c(1000, 500))

#> Total elapsed time =  0.98 seconds 

#> # A tibble: 1,004 x 2
#>    conspecific heterospecific
#>          <dbl>          <dbl>
#>  1        1.23           6.15
#>  2        2              6   
#>  3        2              7   
#>  4        1              9   
#>  5        1              1   
#>  6        1             11   
#>  7        1              5   
#>  8        2              6   
#>  9        1              3   
#> 10        2              7   
#> # ... with 994 more rows

neighbor_basal_area(tree, r = 20, plotdim = c(1000, 500))

#> Total elapsed time =  0.9 seconds 

#> # A tibble: 1,004 x 2
#>    conspecific heterospecific
#>          <dbl>          <dbl>
#>  1      360.             614.
#>  2       14.3            602.
#>  3       17.8            485.
#>  4        8.55          1258.
#>  5      154.             172.
#>  6      475.             726.
#>  7      243.            3180.
#>  8      455.             200.
#>  9     2865.             412.
#> 10       68.5            386.
#> # ... with 994 more rows